+++ modules/tinymce/src/core/demo/ts/demo/ContentEditableFalseDemo.ts
@@ -66,7 +66,7 @@ export default () => {
     skin_url: '../../../../js/tinymce/skins/ui/oxide',
     add_unload_trigger: false,
     toolbar: 'insertfile undo redo | styleselect | bold italic | alignleft aligncenter alignright alignjustify' +
-    ' | bullist numlist outdent indent | link image | print preview media fullpage | forecolor backcolor emoticons table codesample',
+    ' | bullist numlist outdent indent | link image | print preview media | forecolor backcolor emoticons table codesample',
     plugins: [ 'paste', 'list' ],
     content_css: '../css/content_editable.css',
     height: 400
@@ -78,7 +78,7 @@ export default () => {
     skin_url: '../../../../js/tinymce/skins/ui/oxide',
     add_unload_trigger: false,
     toolbar: 'insertfile undo redo | styleselect | bold italic | alignleft aligncenter alignright alignjustify' +
-    ' | bullist numlist outdent indent | link image | print preview media fullpage | forecolor backcolor emoticons table codesample',
+    ' | bullist numlist outdent indent | link image | print preview media | forecolor backcolor emoticons table codesample',
     plugins: [ 'paste', 'list' ],
     content_css: '../css/content_editable.css'
   });