+++ modules/tinymce/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+- Internal selectors were appearing in the style list when using the `importcss` plugin #TINY-8238
+
 ## 5.10.1 - 2021-11-03
 
 ### Fixed